AISIproject/requirements.txt
monai>=1.3.0
timm>=0.9.12
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.22.0
pandas>=2.1.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
tqdm>=4.66.0
pyyaml>=6.0.1
jupyterlab>=4.0.0
